corrected data type bug on bend tag

+++ RocksmithCustomSongCreator/SngFileWriter/SngFileWriter.cs
@@ -802,7 +802,7 @@ namespace RocksmithSngCreator
                     w.Write(notesChords[i].Entity.GetType() == typeof(SongNote) ? ((SongNote)notesChords[i].Entity).Sustain : (Single)0);
 
                     // bend
-                    w.Write(notesChords[i].Entity.GetType() == typeof(SongNote) ? ((SongNote)notesChords[i].Entity).Bend : (Single)0);
+                    w.Write(notesChords[i].Entity.GetType() == typeof(SongNote) ? ((SongNote)notesChords[i].Entity).Bend : (Int32)0);
 
                     // slideTo
                     w.Write(notesChords[i].Entity.GetType() == typeof(SongNote) ? ((SongNote)notesChords[i].Entity).SlideTo : -1);
